requirements.txt
supabase==2.13.0
python-dotenv==1.0.0
google-generativeai>=0.3.2
